Separate permutation module
