+++ moyo/src/base/operation.rs
@@ -49,49 +49,9 @@ pub fn project_rotations(operations: &Operations) -> Rotations {
     operations.iter().map(|ops| ops.rotation).collect()
 }
 
-#[derive(Debug, PartialEq, Eq, Clone, Hash)]
-pub struct Permutation {
-    pub mapping: Vec<usize>,
-}
-
-impl Permutation {
-    pub fn new(mapping: Vec<usize>) -> Self {
-        Self { mapping }
-    }
-
-    pub fn identity(size: usize) -> Self {
-        Self::new((0..size).collect())
-    }
-
-    pub fn size(&self) -> usize {
-        self.mapping.len()
-    }
-
-    pub fn apply(&self, i: usize) -> usize {
-        self.mapping[i]
-    }
-
-    pub fn inverse(&self) -> Self {
-        let mut inv = vec![0; self.size()];
-        for (i, &j) in self.mapping.iter().enumerate() {
-            inv[j] = i;
-        }
-        Self::new(inv)
-    }
-}
-
-impl Mul for Permutation {
-    type Output = Self;
-
-    fn mul(self, rhs: Self) -> Self::Output {
-        let mapping = (0..self.size()).map(|i| self.apply(rhs.apply(i))).collect();
-        Self::new(mapping)
-    }
-}
-
 #[allow(dead_code)]
 /// Used for testing
-pub fn traverse(generators: &Vec<Rotation>) -> Vec<Rotation> {
+pub fn traverse(generators: &Rotations) -> Rotations {
     let mut queue = VecDeque::new();
     let mut visited = HashSet::new();
     let mut group = vec![];
@@ -117,11 +77,9 @@ pub fn traverse(generators: &Vec<Rotation>) -> Vec<Rotation> {
 
 #[cfg(test)]
 mod tests {
-    use std::vec;
-
     use nalgebra::matrix;
 
-    use super::{Permutation, Translation};
+    use super::Translation;
     use crate::base::{lattice::Lattice, Operation};
 
     #[test]
@@ -147,16 +105,5 @@ mod tests {
             0.0, 0.0, 1.0;
         ];
         assert_relative_eq!(actual, expect);
-    }
-
-    #[test]
-    fn test_permutation() {
-        let permutation = Permutation::new(vec![1, 2, 0]);
-        assert_eq!(permutation.apply(0), 1);
-        assert_eq!(permutation.inverse(), Permutation::new(vec![2, 0, 1]));
-        assert_eq!(
-            permutation.clone() * permutation.inverse(),
-            Permutation::identity(3)
-        );
     }
 }
